small adjustments for clearer code to read
I though I was representing max speed wrongly in the upgrade menu because
I noticed I summed 4 to it in the game, hardcoded, and had forgotten why.

Investigating it I found out that 4 is the minimum player speed, so it's
actually correct to sum 4 to the max_speed_limit set by the user in the
upgrade screen. I just turned it into a variable so that in the future,
when I read this code, I can understand what's it about! Also made it
an export var of player in case I feel the need to rebalance this value
in the future.